1. Created User Api in Post man.    (i) Admin can show all list for User.    (ii) admin can show also individual data for user by using user_id.
2. Added Approved_by column in User Table.
   (i) Approved_by column store admin id who is added this user.

3. Added is_published column in:
   (i) size table.
   (ii) color table.
   (iii) brand table .
   (iv) category table.
   (v) seeder.
   *This is a boolean option. admin can activate and deactivate by using is_publish button.

 4. Created a logic for admin. Admin can not    delete anything which is already used.

   (i) Size : admin can not delete size which is already used by a user or admin.
   (ii) color :  admin can not delete color which is already used by a user or admin.
   (iii) brand :  admin can not delete brand which is already used by a user or admin.
   (iv) category :  admin can not delete category which is already used by a user or admin.

+++ app/Http/Controllers/Admin/BrandController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use App\Http\Requests\Brand\StoreBrandRequest;
 use App\Http\Requests\Brand\UpdateBrandRequest;
 use App\Models\Brand;
+use App\Models\Product;
 use Illuminate\Http\Request;
 
 class BrandController extends Controller
@@ -145,6 +146,7 @@ class BrandController extends Controller
         $brand = Brand::create([
             'name' => $brandRequest->name,
             'description' => $brandRequest->description ?? '',
+            'is_published'=>$brandRequest->is_published,
         ]);
 
 //        if ($brandRequest->has('logo')){
@@ -296,6 +298,17 @@ class BrandController extends Controller
     public function destroy(string $id)
     {
         $brand = Brand::find($id);
+
+        if (!$brand) {
+            return response()->json(['success' => false, 'message' => 'Brand not found']);
+        }
+
+        $isBrandUsedInProducts =Product::where('brand_id', $id)->exists();
+
+        if ($isBrandUsedInProducts) {
+            return response()->json(['success' => false, 'message' => 'Brand is in use in product and cannot be deleted'], 403);
+        }
+
         $brand->delete();
         return response()->json(['success' => true, 'message' => 'Brand deleted successfully']);
     }
